+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Clear props before AWQ repopulates the button, patch provided by emmorts on github.
 * Debounce quest list refresh and cache reward preload, patch provided by emmorts on github.
 * Custom Tooltip implementation, patches provided by emmorts on github.
+* Tooltip improvements and refactoring.
 
 # 12.0.1-20260211-1
 * Updated TOC to Midnight Pre-Patch.